Migrate private YEI script to this module
Prior to this, this script was maintained as part of a private Hubot repo
for EverTrue. However, this script is potentially broadly useful.

As a result, I am migrating it here to this repo. This commit represents
the current state of this script.

It should be noted that, as of 2015-10-15, the v1 API that this utilizes
will be shut down. A refactor is currently in-progress.